clean up (removed prints)

+++ Scheduler.h
@@ -12,7 +12,7 @@
 
 class Scheduler {
 public:
-    Scheduler(int numCores, const std::string& type, int timeSlice, float freq, int min, int max, float delay, int memMax, int memFrame, int minMemProc, int maxMemProc);
+    Scheduler(int numCores, const std::string& type, int timeSlice, int freq, int min, int max, int delay, int memMax, int memFrame, int minMemProc, int maxMemProc);
     void addProcess(std::shared_ptr<Process> process);
     void startScheduling();
     void generateProcesses();
